[AOS] refactor: 참여 프로그램 관련 코드 -> 사용자 관련 코드와 통합
