+++ AOS/app/src/main/java/com/hyundai/myexperience/data/entity/user/JoinedProgramList.kt
@@ -1,4 +1,4 @@
-package com.hyundai.myexperience.data.entity.my_page
+package com.hyundai.myexperience.data.entity.user
 
 class JoinedProgramList (
     val joinedPrograms: List<JoinedProgramItem>
