+++ AOS/app/src/main/java/com/hyundai/myexperience/data/remote/service/UserService.kt
@@ -1,6 +1,8 @@
 package com.hyundai.myexperience.data.remote.service
 
-import com.hyundai.myexperience.data.dto.my_page.MyPageResponse
+import com.hyundai.myexperience.data.dto.user.JoinedProgramDetailResponse
+import com.hyundai.myexperience.data.dto.user.JoinedProgramResponse
+import com.hyundai.myexperience.data.dto.user.MyPageResponse
 import com.hyundai.myexperience.data.dto.user.SignInRequest
 import com.hyundai.myexperience.data.dto.user.SignResponse
 import com.hyundai.myexperience.data.dto.user.SignUpRequest
@@ -8,6 +10,8 @@ import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
+import retrofit2.http.Path
+import retrofit2.http.Query
 
 interface UserService {
     @POST("user/login")
@@ -21,4 +25,10 @@ interface UserService {
 
     @GET("user/mypage")
     suspend fun requestMyPage() : Response<MyPageResponse>
+
+    @GET("/user/programs")
+    suspend fun requestJoinedPrograms(@Query("status") status: String): Response<JoinedProgramResponse>
+
+    @GET("/user/participations/{participationId}")
+    suspend fun requestJoinedProgramDetail(@Path("participationId") id: Long): Response<JoinedProgramDetailResponse>
 }
