+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/joined_program/JoinedProgramViewModel.kt
@@ -4,14 +4,15 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hyundai.myexperience.data.repository.JoinedProgramRepository
-import com.hyundai.myexperience.data.entity.my_page.JoinedProgramItem
+import com.hyundai.myexperience.data.entity.user.JoinedProgramItem
+import com.hyundai.myexperience.data.repository.UserRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class JoinedProgramViewModel @Inject constructor(private val repository: JoinedProgramRepository): ViewModel() {
+class JoinedProgramViewModel @Inject constructor(private val repository: UserRepository) :
+    ViewModel() {
     private var _joinedPrograms = MutableLiveData<List<JoinedProgramItem>?>(mutableListOf())
     val joinedPrograms: MutableLiveData<List<JoinedProgramItem>?> = _joinedPrograms
 
@@ -24,8 +25,7 @@ class JoinedProgramViewModel @Inject constructor(private val repository: JoinedP
             if (newJoinedPrograms.isNullOrEmpty()) {
                 _isJoinedProgramsEmpty.value = true
                 _joinedPrograms.value = emptyList()
-            }
-            else {
+            } else {
                 _isJoinedProgramsEmpty.value = false
                 _joinedPrograms.value = newJoinedPrograms
             }
