+++ AOS/app/src/main/java/com/hyundai/myexperience/ui/joined_program/adapter/ProgramsItemClickListener.kt
@@ -1,6 +1,6 @@
 package com.hyundai.myexperience.ui.joined_program.adapter
 
-import com.hyundai.myexperience.data.entity.my_page.JoinedProgramItem
+import com.hyundai.myexperience.data.entity.user.JoinedProgramItem
 
 interface ProgramsItemClickListener {
     fun onItemClick(program : JoinedProgramItem)
